Moved schema into graphql directory
