+++ graphql/schema/schema.js
@@ -1,0 +1,6 @@
+import path from "path";
+import {mergeGraphqlSchemas} from "merge-graphql-schemas";
+
+const schema = mergeGraphqlSchemas(path.join(__dirname, '../'));
+
+export default schema
